Move articles/ to ir/
