+++ components/ir/articles/index.ts
@@ -4,9 +4,9 @@ import {
   addTranche,
   removeTranche,
   updateReformeByName,
-} from "../../redux/actions";
+} from "../../../redux/actions";
 // eslint-disable-next-line no-unused-vars
-import { RootState } from "../../redux/reducers";
+import { RootState } from "../../../redux/reducers";
 import ArticlesComponent from "./articles-component";
 
 const mapStateToProps = ({ parameters }: RootState) => parameters;
