+++ components/ir/articles/quotient-familial/index.js
