+++ components/ir/articles/quotient-familial/regles-generales/QfTable.module.scss
