+++ components/ir/articles/quotient-familial/regles-generales/QfTable.tsx
@@ -6,10 +6,10 @@ import { Fragment, PureComponent } from "react";
 // eslint-disable-next-line no-unused-vars
 import { connect, ConnectedProps } from "react-redux";
 
-import { addNewLineInParameterArray, removeLastLineInParameterArray } from "../../../../redux/actions/parameters";
+import { addNewLineInParameterArray, removeLastLineInParameterArray } from "../../../../../redux/actions/parameters";
 // eslint-disable-next-line no-unused-vars
-import { RootState } from "../../../../redux/reducers";
-import { StateParameter } from "../../../articles-inputs/parameter";
+import { RootState } from "../../../../../redux/reducers";
+import { StateParameter } from "../../../../articles-inputs/parameter";
 import { Button } from "../../buttons";
 import styles from "./QfTable.module.scss";
