+++ components/ir/articles/quotient-familial/regles-specifiques/ReglesSpecifiques.tsx
@@ -1,7 +1,7 @@
 import { Fragment, PureComponent } from "react";
 
-import { StateParameter } from "../../../articles-inputs/parameter";
-import { ExpandablePanelSubTitle, ExpandableText } from "../../../common";
+import { StateParameter } from "../../../../articles-inputs/parameter";
+import { ExpandablePanelSubTitle, ExpandableText } from "../../../../common";
 
 export class ReglesSpecifiques extends PureComponent {
   render() {
