+++ components/ir/index.ts
@@ -1,0 +1,1 @@
+export { default as Articles } from "./articles";
